+++ solana/solana-ibc/programs/solana-ibc/src/lib.rs
@@ -8,39 +8,14 @@ extern crate alloc;
 use anchor_lang::prelude::*;
 use anchor_lang::solana_program;
 use anchor_spl::associated_token::AssociatedToken;
-#[cfg(feature = "mocks")]
-use anchor_spl::token::MintTo;
 use anchor_spl::token::{Mint, Token, TokenAccount};
 use borsh::BorshDeserialize;
-#[cfg(feature = "mocks")]
-use ibc::core::ics03_connection::connection::{
-    ConnectionEnd, Counterparty, State as ConnState,
-};
-#[cfg(feature = "mocks")]
-use ibc::core::ics03_connection::version::Version;
-#[cfg(feature = "mocks")]
-use ibc::core::ics04_channel::channel::{
-    ChannelEnd, Counterparty as ChanCounterparty, Order, State as ChannelState,
-};
-#[cfg(feature = "mocks")]
-use ibc::core::ics04_channel::Version as ChanVersion;
 use ibc::core::ics23_commitment::commitment::CommitmentPrefix;
-#[cfg(feature = "mocks")]
-use ibc::core::ics24_host::identifier::ConnectionId;
 use ibc::core::ics24_host::identifier::{ChannelId, ClientId, PortId};
-#[cfg(feature = "mocks")]
-use ibc::core::ics24_host::path::{
-    ChannelEndPath, ConnectionPath, SeqRecvPath, SeqSendPath,
-};
 use ibc::core::router::{Module, ModuleId, Router};
-#[cfg(feature = "mocks")]
-use ibc::core::ExecutionContext;
 use ibc::core::MsgEnvelope;
 #[cfg(feature = "mocks")]
-use ibc::{
-    core::{ics02_client::ClientExecutionContext, ValidationContext},
-    mock::client_state::MockClientState,
-};
+use mocks::mock_deliver_impl;
 use storage::IbcPackets;
 
 const CHAIN_SEED: &[u8] = b"chain";
@@ -59,12 +34,13 @@ mod error;
 mod events;
 mod execution_context;
 mod host;
+#[cfg(feature = "mocks")]
+mod mocks;
 mod storage;
 #[cfg(test)]
 mod tests;
 mod transfer;
 mod validation_context;
-// mod client_context;
 
 #[anchor_lang::program]
 pub mod solana_ibc {
@@ -204,177 +180,20 @@ pub mod solana_ibc {
     ) -> Result<()> {
         #[cfg(feature = "mocks")]
         {
-            let private: &mut storage::PrivateStorage =
-                &mut ctx.accounts.storage;
-            // msg!("This is private: {private:?}");
-            let provable =
-                storage::get_provable_from(&ctx.accounts.trie, "trie")?;
-            let packets: &mut IbcPackets = &mut ctx.accounts.packets;
-            let accounts = ctx.remaining_accounts;
-
-            let host_head = host::Head::get()?;
-            let (host_timestamp, host_height) = host_head
-                .ibc_timestamp()
-                .and_then(|ts| host_head.ibc_height().map(|h| (ts, h)))
-                .map_err(error::Error::from)
-                .map_err(|err| error!((&err)))?;
-
-            let mut store =
-                storage::IbcStorage::new(storage::IbcStorageInner {
-                    private,
-                    provable,
-                    packets,
-                    accounts: accounts.to_vec(),
-                    host_head,
-                });
-
-            let any_client_state = store.client_state(&client_id).unwrap();
-            let client_state: MockClientState = match any_client_state {
-                client_state::AnyClientState::Tendermint(_) => {
-                    panic!("Only for mocks")
-                }
-                client_state::AnyClientState::Mock(mock) => mock,
-            };
-
-
-            // Store update time since its not called during mocks
-            store
-                .store_update_time(
-                    client_id.clone(),
-                    client_state.latest_height(),
-                    host_timestamp,
-                )
-                .unwrap();
-            store
-                .store_update_height(
-                    client_id.clone(),
-                    client_state.latest_height(),
-                    host_height,
-                )
-                .unwrap();
-
-            let connection_id_on_a = ConnectionId::new(0);
-            let connection_id_on_b = ConnectionId::new(1);
-            let delay_period = core::time::Duration::from_nanos(0);
-            let connection_counterparty = Counterparty::new(
-                counterparty_client_id.clone(),
-                Some(connection_id_on_b.clone()),
+            mock_deliver_impl(
+                ctx,
+                port_id,
+                channel_id,
+                base_denom,
                 commitment_prefix,
-            );
-            let connection_end_on_a = ConnectionEnd::new(
-                ConnState::Open,
-                client_id.clone(),
-                connection_counterparty.clone(),
-                vec![Version::default()],
-                delay_period,
-            )
-            .unwrap();
-            let connection_end_on_b = ConnectionEnd::new(
-                ConnState::Open,
                 client_id,
-                connection_counterparty,
-                vec![Version::default()],
-                delay_period,
+                counterparty_client_id,
             )
             .unwrap();
-
-            let counterparty =
-                ChanCounterparty::new(port_id.clone(), Some(ChannelId::new(0)));
-            let channel_end_on_a = ChannelEnd::new(
-                ChannelState::Open,
-                Order::Unordered,
-                counterparty.clone(),
-                vec![connection_id_on_a.clone()],
-                ChanVersion::new(
-                    ibc::applications::transfer::VERSION.to_string(),
-                ),
-            )
-            .unwrap();
-            let channel_end_on_b = ChannelEnd::new(
-                ChannelState::Open,
-                Order::Unordered,
-                counterparty,
-                vec![connection_id_on_b.clone()],
-                ChanVersion::new(
-                    ibc::applications::transfer::VERSION.to_string(),
-                ),
-            )
-            .unwrap();
-            let channel_id_on_a = ChannelId::new(0);
-            let channel_id_on_b = ChannelId::new(1);
-
-            // For Client on Chain A
-            store
-                .store_connection(
-                    &ConnectionPath(connection_id_on_a),
-                    connection_end_on_a,
-                )
-                .unwrap();
-            store
-                .store_channel(
-                    &ChannelEndPath(port_id.clone(), channel_id_on_a.clone()),
-                    channel_end_on_a,
-                )
-                .unwrap();
-            store
-                .store_next_sequence_send(
-                    &SeqSendPath(port_id.clone(), channel_id_on_a.clone()),
-                    1.into(),
-                )
-                .unwrap();
-            store
-                .store_next_sequence_recv(
-                    &SeqRecvPath(port_id.clone(), channel_id_on_a),
-                    1.into(),
-                )
-                .unwrap();
-
-            // For Client on chain b
-            store
-                .store_connection(
-                    &ConnectionPath(connection_id_on_b),
-                    connection_end_on_b,
-                )
-                .unwrap();
-            store
-                .store_channel(
-                    &ChannelEndPath(port_id.clone(), channel_id_on_b.clone()),
-                    channel_end_on_b,
-                )
-                .unwrap();
-            store
-                .store_next_sequence_send(
-                    &SeqSendPath(port_id.clone(), channel_id_on_b.clone()),
-                    1.into(),
-                )
-                .unwrap();
-            store
-                .store_next_sequence_recv(
-                    &SeqRecvPath(port_id, channel_id_on_b),
-                    1.into(),
-                )
-                .unwrap();
-
-            // Minting some tokens to the authority so that he can do the transfer
-            let bump_vector =
-                ctx.bumps.get("mint_authority").unwrap().to_le_bytes();
-            let inner = vec![MINT_ESCROW_SEED, bump_vector.as_ref()];
-            let outer = vec![inner.as_slice()];
-
-            // Mint some tokens to escrow account
-            let mint_instruction = MintTo {
-                mint: ctx.accounts.token_mint.to_account_info(),
-                to: ctx.accounts.sender_token_account.to_account_info(),
-                authority: ctx.accounts.mint_authority.to_account_info(),
-            };
-            let cpi_ctx = CpiContext::new_with_signer(
-                ctx.accounts.token_program.to_account_info(),
-                mint_instruction,
-                outer.as_slice(), //signer PDA
-            );
-            anchor_spl::token::mint_to(cpi_ctx, 10000000)?;
+            Ok(())
         }
-        Ok(())
+        #[cfg(not(feature = "mocks"))]
+        panic!("This method is only for mocks");
     }
 }
